+++ src/cone/app/browser/exception.py
@@ -1,15 +1,18 @@
+from cone.app.browser import render_main_template
+from cone.app.browser.login import login_view
+from cone.app.browser.utils import format_traceback
+from cone.tile import registerTile
+from pyramid.httpexceptions import HTTPForbidden
+from pyramid.httpexceptions import HTTPNotFound
 from pyramid.response import Response
+from pyramid.security import authenticated_userid
 from pyramid.view import view_config
 import json
-import sys
-import traceback
 
 
-def format_traceback():
-    etype, value, tb = sys.exc_info()
-    ret = ''.join(traceback.format_exception(etype, value, tb))
-    return '<pre>%s</pre>' % ret
-
+###############################################################################
+# Internal Server Error
+###############################################################################
 
 ERROR_PAGE = """
 <h1>An error occured</h1>
@@ -40,3 +43,31 @@ def internal_server_error(request):
     response = Response(json.dumps(ret))
     response.content_type = 'application/json'
     return response
+
+
+###############################################################################
+# Unauthorized
+###############################################################################
+
+registerTile('unauthorized', 'templates/unauthorized.pt', permission='login')
+
+
+@view_config(context=HTTPForbidden)
+def forbidden_view(request):
+    model = request.context
+    if not authenticated_userid(request):
+        return login_view(model, request)
+    return render_main_template(model, request, contenttile='unauthorized')
+
+
+###############################################################################
+# Not Found
+###############################################################################
+
+registerTile('not_found', 'templates/not_found.pt', permission='login')
+
+
+@view_config(context=HTTPNotFound)
+def not_found_view(request):
+    model = request.context
+    return render_main_template(model, request, contenttile='not_found')